pyproject.toml
[tool]
[tool.poetry]
authors = ["Ben Pedigo <ben.pedigo@alleninstitute.org>"]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.11',
]
description = "Skeletonization service for the CAVE ecosystem."
homepage = "https://github.com/bdpedigo/skervice"
license = "MIT"
name = "skervice"
packages = [{ include = "skervice" }, { include = "tests", format = "sdist" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
beartype = "^0.16.4"
brotlicffi = "^1.1.0.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"
numpy = "^1.26.3"
pandas = "^2.1.4"
python = ">=3.11.0,<3.12.0"
messagingclient = {git = "https://github.com/seung-lab/messaging-client.git"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
commitizen = "^3.13.0"
mike = "^2.0.0"
mkdocs = "^1.5.3"
mkdocs-autorefs = "^0.5.0"
mkdocs-git-authors-plugin = "^0.7.2"
mkdocs-git-committers-plugin-2 = "^2.2.2"
mkdocs-git-revision-date-localized-plugin = "^1.2.2"
mkdocs-include-markdown-plugin = "^6.0.4"
mkdocs-jupyter = "^0.24.6"
mkdocs-material = "^9.5.2"
mkdocs-material-extensions = "^1.3.1"
mkdocs-section-index = "^0.3.8"
mkdocstrings = { extras = ["crystal", "python"], version = "^0.24.0" }
mknotebooks = "^0.8.0"
mypy = "^1.7.1"
pymdown-extensions = "^10.5"
pytest = "^7.4.3"
ruff = "^0.1.7"
tox = "^4.11.4"
tox-current-env = "^0.0.11"
twine = "^4.0.2"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry"]

[tool.ruff]
ignore-init-module-imports = true
